fix(discord): suppress #3964 footer on fresh synthetic turns via completion_background + fit namespace cap
Adjudicated review of #3967 found the first pass missed the COMMON case: a fresh
`/loop`·MonitorAutoTurn (or background task-notification) turn. Root of the miss
(confirmed in tmux_watcher.rs): `turn_is_external_input_for_session` is computed
ONCE at the top of each `'watcher_loop` iteration (`:1017`) from the `:1009`
inflight snapshot, but a fresh synthetic turn's row is created LATER in the SAME
iteration (`ensure_monitor_auto_turn_inflight`, `:1226`/`:1495`). When that turn
emits `Done`, the terminal chokepoint (`:5704`) runs in the same iteration with the
cached flag still `false` (the `:2013` late set-true is in the separate-panel
branch, mutually exclusive with footer mode), so the first gate
`(footer_mode && turn_is_external_input_for_session)` did not fire.

Fix: widen the gate with `completion_background`
(`task_notification_kind ∈ {Background, MonitorAutoTurn}`), which is computed at the
chokepoint itself (`:5700`) from the STREAM-derived `task_notification_kind`
(merged at `:1170`/`:1464`), so it is correct regardless of inflight-row timing:

    footer_mode && (turn_is_external_input_for_session || completion_background)

`completion_background` is ALREADY a parameter of
`complete_watcher_terminal_footer_or_status_panel`, so the fix stays entirely in the
non-hotfile `tmux_watcher/` submodules — tmux_watcher.rs and every #3016 hotfile
remain untouched. A genuine Discord-origin turn carries no `<task-notification>`
marker (`completion_background = false`) and is not panel-eligible, so its #3089
footer is preserved.

Namespace-cap compliance (audit_maintainability hard gate, 700 LoC for
`tmux_watcher/**`): the first pass pushed single_message_footer.rs to 830 LoC — over
cap (the earlier `| tail` masked the real exit code). Relocate the two PURE
decisions — the suppression gate and `watcher_completion_footer_should_tick` — plus
their tests to the sibling `panel_decisions.rs` (their natural home; pure watcher
decisions, alongside `watcher_inflight_is_panel_eligible_for_session`). Result:
single_message_footer.rs 684 LoC, panel_decisions.rs 643 LoC — both under cap.

Tests (panel_decisions.rs::completion_footer_suppression_tests):
- gate_suppresses_for_either_mirror_signal_and_keeps_discord_origin_3964 — pins the
  fresh-synthetic regression (cached flag false + completion_background true → still
  suppressed) AND #3089 (both false → footer kept).
- synthetic_tui_mirror_emits_prose_without_tui_chrome_3964
- synthetic_tui_mirror_finalize_strips_live_status_footer_to_prose_3964

Gates: fmt --check, audit_maintainability --check, check_hotfile_ratchet all exit 0;
cargo test --lib (panel_decisions + both single_message_footer modules) green incl.
the #3959 bridge tests; inventory + agent-maintenance freshness passed.
